+++ gkraken/app.py
@@ -29,7 +29,7 @@ from gkraken.conf import APP_NAME, APP_ID, APP_VERSION
 from gkraken.di import MainBuilder
 from gkraken.model import SpeedProfile, SpeedStep, Setting, CurrentSpeedProfile, load_db_default_data
 from gkraken.presenter.main import MainPresenter
-from gkraken.util.autostart import set_autostart_enabled
+from gkraken.util.desktop_entry import set_autostart_entry, add_application_entry
 from gkraken.util.log import LOG_DEBUG_FORMAT
 from gkraken.util.udev import add_udev_rule, remove_udev_rule
 from gkraken.util.view import build_glib_option
@@ -105,6 +105,21 @@ class Application(Gtk.Application):
             LOG.debug("Option %s selected", _Options.HIDE_WINDOW.value)
             self._start_hidden = True
 
+        if _Options.APPLICATION_ENTRY.value in options:
+            LOG.debug("Option %s selected", _Options.APPLICATION_ENTRY.value)
+            add_application_entry()
+            start_app = False
+
+        if _Options.AUTOSTART_ON.value in options:
+            LOG.debug("Option %s selected", _Options.AUTOSTART_ON.value)
+            set_autostart_entry(True)
+            start_app = False
+
+        if _Options.AUTOSTART_OFF.value in options:
+            LOG.debug("Option %s selected", _Options.AUTOSTART_OFF.value)
+            set_autostart_entry(True)
+            start_app = False
+
         if _Options.UDEV_ADD_RULE.value in options:
             LOG.debug("Option %s selected", _Options.UDEV_ADD_RULE.value)
             exit_value += add_udev_rule()
@@ -113,16 +128,6 @@ class Application(Gtk.Application):
         if _Options.UDEV_REMOVE_RULE.value in options:
             LOG.debug("Option %s selected", _Options.UDEV_REMOVE_RULE.value)
             exit_value += remove_udev_rule()
-            start_app = False
-
-        if _Options.AUTOSTART_ON.value in options:
-            LOG.debug("Option %s selected", _Options.AUTOSTART_ON.value)
-            set_autostart_enabled(True)
-            start_app = False
-
-        if _Options.AUTOSTART_OFF.value in options:
-            LOG.debug("Option %s selected", _Options.AUTOSTART_OFF.value)
-            set_autostart_enabled(True)
             start_app = False
 
         if start_app:
@@ -141,6 +146,8 @@ class Application(Gtk.Application):
                               description="Start with the main window hidden"),
         ]
         linux_options = [
+            build_glib_option(_Options.APPLICATION_ENTRY.value,
+                              description="Add a desktop entry for the application"),
             build_glib_option(_Options.AUTOSTART_ON.value,
                               description="Enable automatic start of the app on login"),
             build_glib_option(_Options.AUTOSTART_OFF.value,
@@ -161,6 +168,7 @@ class _Options(Enum):
     VERSION = 'version'
     DEBUG = 'debug'
     HIDE_WINDOW = 'hide-window'
+    APPLICATION_ENTRY = 'application-entry'
     AUTOSTART_ON = 'autostart-on'
     AUTOSTART_OFF = 'autostart-off'
     UDEV_ADD_RULE = 'udev-add-rule'